+++ src/main/frontend/common/components/status-icon.scss
