+++ src/main/frontend/common/utils/timings.ts
